requirements.txt
anyascii==0.3.1
appdirs==1.4.4
audioread==3.0.0
autopep8==2.0.0
Babel==2.11.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
coqpit==0.0.9
coqui-stt-model-manager==0.0.21
cycler==0.11.0
Cython==0.29.28
dateparser==1.1.4
decorator==5.1.1
docopt==0.6.2
et-xmlfile==1.1.0
evdev==1.6.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SocketIO==4.3.2
fonttools==4.38.0
fsspec==2022.11.0
g2pkk==0.1.2
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
halo==0.0.31
idna==2.10
inflect==5.6.0
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.0.1
joblib==1.2.0
jsonlines==1.2.0
kiwisolver==1.4.4
librosa==0.8.0
llvmlite==0.38.1
log-symbols==0.0.14
MarkupSafe==2.1.1
matplotlib==3.6.2
mecab-python3==1.0.5
networkx==2.8.8
nltk==3.7
num2words==0.5.12
numba==0.55.2
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai==0.25.0
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pandas-stubs==1.5.2.221124
Pillow==9.3.0
pooch==1.6.0
protobuf==3.19.6
psutil==5.9.4
PyAudio==0.2.12
pycodestyle==2.10.0
pycparser==2.21
pynndescent==0.5.8
pynput==1.7.6
pyparsing==3.0.9
pypinyin==0.47.1
pysbd==0.3.4
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==3.14.2
python-socketio==4.6.1
python-xlib==0.32
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==6.0
regex==2022.10.31
requests==2.25.1
resampy==0.4.2
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
soundfile==0.11.0
spinners==0.0.24
stt==1.4.0
tensorboardX==2.5.1
termcolor==2.1.1
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
tqdm==4.64.1
trainer==0.0.17
TTS==0.9.0
types-pytz==2022.6.0.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
umap-learn==0.5.1
unidic-lite==1.0.8
urllib3==1.26.13
webrtcvad==2.0.10
Werkzeug==2.0.3
